Merge phieu tinh lai

+++ DTO/PhieuGoiTien.cs
@@ -12,18 +12,18 @@ namespace DTO
         public string SoTaiKhoanThanhToan { get; set; }
         public string TenKhachHang { get; set; }
         public string MaLoaiTietKiem { get; set; }
-        public float LaiSuatApDung { get; set; }
-        public float LaiSuatPhatSinh { get; set; }
+        public double LaiSuatApDung { get; set; }
+        public double LaiSuatPhatSinh { get; set; }
         public DateTime NgayGoi { get; set; }
         public DateTime NgayDaoHanKeTiep { get; set; }
-        public float TongTienGoc { get; set; }
-        public float TongTienLaiPhatSinh { get; set; }
+        public double TongTienGoc { get; set; }
+        public double TongTienLaiPhatSinh { get; set; }
         public int HinhThucGiaHan { get; set; }
         public PhieuGoiTien() { }
 
         // Constructor
-        public PhieuGoiTien(string soTaiKhoanTienGoi, string soTaiKhoanThanhToan, string tenKH, string maLoaiTietKiem, float laiSuatApDung, float laiSuatPhatSinh,
-                            DateTime ngayGoi, DateTime ngayDaoHanKeTiep, float tongTienGoc, float tongTienLaiPhatSinh, int hinhThucGiaHan)
+        public PhieuGoiTien(string soTaiKhoanTienGoi, string soTaiKhoanThanhToan, string tenKH, string maLoaiTietKiem, double laiSuatApDung, double laiSuatPhatSinh,
+                            DateTime ngayGoi, DateTime ngayDaoHanKeTiep, double tongTienGoc, double tongTienLaiPhatSinh, int hinhThucGiaHan)
         {
             SoTaiKhoanTienGoi = soTaiKhoanTienGoi;
             SoTaiKhoanThanhToan = soTaiKhoanThanhToan;
